requirements.txt
absl-py==2.1.0
astunparse==1.6.3
boto3==1.35.34
botocore==1.35.34
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
flatbuffers==25.1.21
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imutils==0.5.4
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==10.4.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.5.0
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.14.1
